requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
jinja2
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pillow
paho-mqtt
requests
opencv-python
easyocr
